chore: move embed to a proper pkg/
