+++ maps/tiles.go
@@ -5,7 +5,7 @@ import (
 	"image"
 	"image/draw"
 
-	"github.com/zwindler/gocastle/utils"
+	"github.com/zwindler/gocastle/pkg/embed"
 
 	_ "image/jpeg" // Import image/jpeg to support JPEG images
 	_ "image/png"  // Import image/png to support PNG images
@@ -73,7 +73,7 @@ var (
 
 // extractTileFromTileset extracts a subimage from coordinates on a tileset.
 func extractTileFromTileset(x, y int, filePath string) (image.Image, error) {
-	file, err := utils.EmbeddedImages.Open(filePath)
+	file, err := embed.EmbeddedImages.Open(filePath)
 	if err != nil {
 		fmt.Println("Error opening image:", err)
 		return nil, err
