+++ pkg/embed/embed.go
@@ -1,4 +1,4 @@
-package utils
+package embed
 
 import (
 	"embed"
