+++ pkg/embed/static/RPG Nature Tileset.png
