+++ pkg/embed/static/bald_man.png
