+++ pkg/embed/static/bald_person.png
