+++ pkg/embed/static/bald_woman.png
