+++ pkg/embed/static/blond_haired_man.png
