+++ pkg/embed/static/blond_haired_person.png
