+++ pkg/embed/static/blond_haired_woman.png
