+++ pkg/embed/static/castle_back.png
