+++ pkg/embed/static/chest.png
