+++ pkg/embed/static/dark_haired_man.png
