+++ pkg/embed/static/female-farmer.png
