+++ pkg/embed/static/giant-ant.png
