+++ pkg/embed/static/giant-red-ant.png
