+++ pkg/embed/static/goblin-short.png
