+++ pkg/embed/static/goblin.png
