+++ pkg/embed/static/inventory.png
