+++ pkg/embed/static/kobold-short.png
