+++ pkg/embed/static/kobold.png
