+++ pkg/embed/static/male-farmer.png
