+++ pkg/embed/static/minotaur-short.png
