+++ pkg/embed/static/minotaur.png
