+++ pkg/embed/static/ogre.png
