+++ pkg/embed/static/red_haired_man.png
