+++ pkg/embed/static/red_haired_person.png
