+++ pkg/embed/static/red_haired_woman.png
