+++ pkg/embed/static/scarf_woman.png
