+++ pkg/embed/static/sword.png
