+++ pkg/embed/static/tilea1_MACK.png
