+++ pkg/embed/static/tilea2_MACK.png
