+++ pkg/embed/static/tilea3_MACK.png
