+++ pkg/embed/static/tilea4_MACK.png
