+++ pkg/embed/static/tilea5_MACK.png
