+++ pkg/embed/static/tileb_MACK.png
