+++ pkg/embed/static/tilec_MACK.png
