+++ pkg/embed/static/tiled_MACK.png
