+++ pkg/embed/static/tilee_MACK.png
