+++ pkg/embed/static/transparent_tile.png
