+++ pkg/embed/static/white_haired_man.png
