+++ pkg/embed/static/white_haired_person.png
