+++ pkg/embed/static/white_haired_woman.png
