+++ pkg/embed/static/wolf.png
