+++ pkg/embed/static/woman-mage.png
